fix: los 4 technische issues op (#88, #75, #80, #86)
- #75: verplaats 4 ongebruikte functies uit transforms.py naar archive/transforms_unused.py
- #80: vervang collections.Counter-intersectie door set-lookup (base.py, individual.py, weeks.py)
- #86: toon expliciete waarschuwing bij stille downgrade combined → cumulative
- #88: voeg docstring toe aan BaseForecaster ABC met motivatie voor de abstractie
