+++ src/studentprognose/models/base.py
@@ -3,7 +3,12 @@ import numpy as np
 
 
 class BaseForecaster(ABC):
-    """Base class for time-series forecasting models."""
+    """Abstracte basisklasse voor tijdreeksmodellen.
+
+    Momenteel is `SARIMAForecaster` de enige implementatie. De abstractie is
+    intentioneel: toekomstige modellen (bijv. Prophet, ETS) kunnen dezelfde
+    interface implementeren zodat ze uitwisselbaar zijn in de ensemble-pipeline.
+    """
 
     @abstractmethod
     def fit(self, ts_data: np.ndarray, exog: np.ndarray | None = None) -> "BaseForecaster":
